Unstable commit - Firebase integration Integration of Firebase database in proress. Decided to use Firestore oer realtime due to prior experience with firestore. 1. Updated build related dependencies in gradle for firease integration. 2. Sample database created on firestore website manually and connected it with the app in android studio. 3. Able to read sample information present in database in the app using the Firebase related classes. This part is under progress. 4. No authentication is set up in firebase directly. User varification will be done manually by storing the username-password combos in the repo itself.
Read and write using Firecloud to be explored more. Post this, proper class definitions should be made for storing information in sync with the storage structure followed in firebase.
Eventually, all download from firebase will happen at once and not everytime a screen is opened newly. This will save unnecessary re-download of the same information again and again.